got rid of circular dependencies
