+++ packages/bluerain-os/src/plugins/defaultPlugins.js
@@ -1,10 +1,10 @@
 // This file registers builtin plugins that ship with BlueRain OS
-import ReduxPlugin from '@blueeast/bluerain-plugin-redux';
 import ReactRouterPlugin from '@blueeast/bluerain-plugin-react-router';
+import ReduxPlugin from '@blueeast/bluerain-plugin-redux';
 import WindowInfoPlugin from './WindowInfoPlugin';
 
 export default [
-	ReduxPlugin,
 	ReactRouterPlugin,
+	ReduxPlugin,
 	WindowInfoPlugin
 ];
