+++ packages/bluerain-plugin-intl/package.json
@@ -13,7 +13,8 @@
   "author": "BlueEast Team",
   "license": "ISC",
   "dependencies": {
-    "@blueeast/bluerain-os": "git://github.com/BlueEastCode/bluerain-os.git#next",
+    "@blueeast/bluerain-os": "^0.1.0",
+    "react": "^16.0.0-alpha.12",
     "react-intl": "^2.4.0"
   },
   "devDependencies": {
