+++ packages/bluerain-plugin-react-router/package.json
@@ -13,10 +13,11 @@
   "author": "BlueEast Team",
   "license": "ISC",
   "dependencies": {
-    "@blueeast/bluerain-os": "git://github.com/BlueEastCode/bluerain-os.git#next",
-		"react-router-dom": "^4.2.2",
-		"react-router-native": "^4.2.0",
-		"react-router-redux": "^4.0.8"
+    "@blueeast/bluerain-os": "^0.1.0",
+    "react": "^16.0.0-alpha.12",
+    "react-router-dom": "^4.2.2",
+    "react-router-native": "^4.2.0",
+    "react-router-redux": "^5.0.0-alpha.6"
   },
   "devDependencies": {
     "babel-cli": "6.16.0",
