+++ packages/bluerain-plugin-react-router/src/index.js
@@ -1,3 +1,5 @@
+import Plugin from './Plugin';
+
 export * from 'react-router-dom';
 export { ConnectedRouter as SystemRouter } from 'react-router-redux';
-export { Plugin as default } from './Plugin';
+export default Plugin;
