+++ packages/bluerain-plugin-redux/package.json
@@ -13,7 +13,8 @@
   "author": "BlueEast Team",
   "license": "ISC",
   "dependencies": {
-    "@blueeast/bluerain-os": "git://github.com/BlueEastCode/bluerain-os.git#next",
+    "@blueeast/bluerain-os": "^0.1.0",
+    "react": "^16.0.0-alpha.12",
     "react-redux": "^5.0.6",
     "redux": "^3.7.2"
   },
