+++ packages/demo-electron/package.json
@@ -10,7 +10,7 @@
     "lint": "./node_modules/.bin/eslint ./src/"
   },
   "dependencies": {
-    "@blueeast/bluerain-os": "git://github.com/BlueEastCode/bluerain-os.git#next",
+    "@blueeast/bluerain-os": "^0.1.0",
     "react": "16.0.0-alpha.12",
     "react-dom": "^16.0.0-alpha.12",
     "react-native": "0.46.4",
