+++ packages/demo-web/package.json
@@ -9,11 +9,11 @@
     "start": "webpack-dev-server --inline --hot --history-api-fallback"
   },
   "dependencies": {
-    "@blueeast/bluerain-app-hello-world": "0.0.0",
-    "@blueeast/bluerain-os": "0.1.0",
-    "@blueeast/bluerain-plugin-apollo": "0.1.0",
-    "@blueeast/bluerain-plugin-intl": "0.1.0",
-    "@blueeast/bluerain-plugin-redux-devtools": "0.1.0",
+    "@blueeast/bluerain-app-hello-world": "^0.0.0",
+    "@blueeast/bluerain-os": "^0.1.0",
+    "@blueeast/bluerain-plugin-apollo": "^0.1.0",
+    "@blueeast/bluerain-plugin-intl": "^0.1.0",
+    "@blueeast/bluerain-plugin-redux-devtools": "^0.1.0",
     "react": "^16.0.0-alpha.12",
     "react-dom": "^16.0.0-alpha.12",
     "reactxp": "^0.42.0-rc.25"
